v-dukut/requirements.txt
pyodbc
azure-kusto-data
pandas
numpy